Added setting for item page cache, moved mixin to views utils
