+++ python/aristotle-metadata-registry/aristotle_mdr/required_settings.py
@@ -269,3 +269,6 @@ AUTH_PASSWORD_VALIDATORS = [
 
 # GeoIP
 GEOIP_PATH = os.path.join(BASE_DIR, 'aristotle_mdr/vendor/geoip')
+
+# Caching
+CACHE_ITEM_PAGE = True
